refactor(iterate): split oversize iterate-entry encoding test file
PR #565 grew test_iterate_entry_encoding.py to 318 lines while adding
tail-fallback fail-closed coverage, crossing the 300-line bloat
guideline as a new crossing (flagged by the Stop hook after that PR
had already merged).

Splits it into itself (161 lines, the original decode-safety scope,
trg-06216b9f) and a new sibling test_iterate_entry_tail_fallback.py
(172 lines, the tail-fallback fail-closed fix, trg-e0a0f569). No test
logic changed -- identical outcome before and after (105 passed, 1
skipped).

Run-ID: iterate-2026-08-05-split-iterate-entry-tail-fallback-tests
