+++ .shipwright/agent_docs/iterates/iterate-2026-08-05-split-iterate-entry-tail-fallback-tests.json
@@ -1,0 +1,11 @@
+{
+  "run_id": "iterate-2026-08-05-split-iterate-entry-tail-fallback-tests",
+  "date": "2026-08-05T20:36:29.872130Z",
+  "event_at": "2026-08-05T20:12:35.724764+00:00",
+  "type": "change",
+  "complexity": "trivial",
+  "branch": "iterate/split-iterate-entry-tail-fallback-tests",
+  "spec": null,
+  "tests_passed": true,
+  "adr": "iterate-2026-08-05-split-iterate-entry-tail-fallback-tests"
+}
