+++ CHANGELOG-unreleased.d/Changed/iterate-2026-08-05-split-iterate-entry-tail-fallback-tests_001.md
@@ -1,0 +1,1 @@
+split the oversized test_iterate_entry_encoding.py into itself plus a new test_iterate_entry_tail_fallback.py, no behavior change
